move emacs modules
