+++ src/Juvix/Compiler/Nockma/Highlight/Base.hs
@@ -1,6 +1,6 @@
 module Juvix.Compiler.Nockma.Highlight.Base
-  ( module Juvix.Compiler.Concrete.Data.Highlight.Properties,
+  ( module Juvix.Emacs.Properties,
   )
 where
 
-import Juvix.Compiler.Concrete.Data.Highlight.Properties
+import Juvix.Emacs.Properties
